feat(metrics): float anomaly path for fractional metrics (tripl-68bc)
The detector consumed integer SeriesPoints, so sub-unit ratio/avg/
per_distinct_user movements rounded toward 0 and the 1.0 absolute stddev
floor flattened any surviving delta to z~0 — real fractional spikes were
undetectable. Now:
- SeriesPoint/DetectedAnomaly carry floats; catalog metric series load
  unrounded.
- Fractional series (fill_gaps=False) derive their absolute stddev floor
  from the series' own robust magnitude (1% of median |value|), so a
  0.5 -> 0.8 ratio move scores as the multi-sigma event it is while
  sub-1% wobble stays suppressed.
- actual_count columns go BigInteger->Float on both anomaly tables, and the
  alert-delivery items keep unrounded actual/expected/delta (migration
  f9a0b1c2d3e4); alert text renders floats via the shared stringifier
  (integral values keep printing as plain ints).
- Response schemas + regenerated OpenAPI/types expose float actuals.

+++ backend/src/tripl/anomaly_context.py
@@ -48,7 +48,7 @@ def build_sparkline(values: Iterable[float | int], width: int = SPARKLINE_WIDTH)
     return "".join(chars)
 
 
-def _percent_signed(actual: int, expected: float) -> str:
+def _percent_signed(actual: float, expected: float) -> str:
     """+320% / -75% / inf% style label, always rounded to integer."""
     if expected <= 0:
         return "+inf%" if actual > 0 else "0%"
